Update offset & obsolete field names

+++ Craftimizer/Utils/CSCraftEventHandler.cs
@@ -7,7 +7,7 @@ namespace Craftimizer.Utils;
 public unsafe struct CSCraftEventHandler
 {
     [FieldOffset(0x48A)] public unsafe fixed ushort WKSClassLevels[2];
-    [FieldOffset(0x48E)] public unsafe fixed ushort WKSClassJobs[2];
+    [FieldOffset(0x48E)] public unsafe fixed byte WKSClassJobs[2];
 
     public static CSCraftEventHandler* Instance()
     {
